+++ svd-parser/src/cluster.rs
@@ -25,10 +25,9 @@ impl Parse for Cluster {
                     .at(tree.id()));
                 }
             }
-
-            Ok(Self::Array(info, array_info))
+            Ok(info.array(array_info))
         } else {
-            Ok(Self::Single(info))
+            Ok(info.single())
         }
     }
 }